move positioners to their own file
